Remove warnings

+++ src/protocols/ProtocolMqtt5Client.cpp
@@ -20,15 +20,14 @@
 //OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //SOFTWARE.
 
-
 #include "finalmq/protocols/ProtocolMqtt5Client.h"
-#include "finalmq/streamconnection/Socket.h"
-#include "finalmq/protocolsession/ProtocolMessage.h"
-#include "finalmq/protocolsession/ProtocolRegistry.h"
-#include "finalmq/variant/VariantValueStruct.h"
-#include "finalmq/variant/VariantValues.h"
 
 #include "finalmq/helpers/ModulenameFinalmq.h"
+#include "finalmq/protocolsession/ProtocolMessage.h"
+#include "finalmq/protocolsession/ProtocolRegistry.h"
+#include "finalmq/streamconnection/Socket.h"
+#include "finalmq/variant/VariantValueStruct.h"
+#include "finalmq/variant/VariantValues.h"
 
 #ifdef WIN32
 #include <rpcdce.h>
@@ -46,10 +45,8 @@
 #endif
 #endif
 
-
-
-namespace finalmq {
-
+namespace finalmq
+{
 const std::uint32_t ProtocolMqtt5Client::PROTOCOL_ID = 5;
 const std::string ProtocolMqtt5Client::PROTOCOL_NAME = "mqtt5client";
 
@@ -69,7 +66,6 @@ static const std::string TOPIC_WILLMESSAGE = "/fmq_willmsg";
 static const std::string SESSIONID_PREFIX = "/_id:";
 
 static const std::uint8_t ReasonCodeDisconnectWithWillMessage = 0x04;
-
 
 #ifdef WIN32
 static std::string getUuid()
@@ -141,7 +137,7 @@ void uuid_unparse(uuid_t uu, char* out)
 static std::string getUuid()
 {
     std::string strUuid;
-    strUuid.resize(50);     // 37 is the exact number
+    strUuid.resize(50); // 37 is the exact number
     uuid_t uuid;
     uuid_generate(uuid);
     uuid_unparse(uuid, const_cast<char*>(strUuid.data()));
@@ -152,7 +148,7 @@ static std::string getUuid()
 static std::string getUuid()
 {
     std::string strUuid;
-    strUuid.resize(50);     // 37 is the exact number
+    strUuid.resize(50); // 37 is the exact number
     uuid_t uuid;
     uuid_generate(uuid);
     uuid_unparse(uuid, const_cast<char*>(strUuid.data()));
@@ -161,8 +157,6 @@ static std::string getUuid()
 }
 #endif
 #endif
-
-
 
 ProtocolMqtt5Client::ProtocolMqtt5Client(const Variant& data)
     : m_client(std::make_unique<Mqtt5Client>())
@@ -185,8 +179,6 @@ ProtocolMqtt5Client::ProtocolMqtt5Client(const Variant& data)
     m_client->setCallback(this);
 }
 
-
-
 ProtocolMqtt5Client::~ProtocolMqtt5Client()
 {
     if (m_connection)
@@ -194,7 +186,6 @@ ProtocolMqtt5Client::~ProtocolMqtt5Client()
         m_connection->disconnect();
     }
 }
-
 
 // IProtocol
 void ProtocolMqtt5Client::setCallback(const std::weak_ptr<IProtocolCallback>& callback)
@@ -211,22 +202,21 @@ void ProtocolMqtt5Client::setConnection(const IStreamConnectionPtr& connection)
     data.receiveMaximum = 128;
     data.username = m_username;
     data.password = m_password;
-    data.keepAlive = m_keepAlive;
+    data.keepAlive = static_cast<std::uint16_t>(m_keepAlive);
     data.sessionExpiryInterval = m_sessionExpiryInterval;
     data.willMessage = std::make_unique<IMqtt5Client::WillMessage>();
     data.willMessage->payload.insert(data.willMessage->payload.end(), m_virtualSessionId.begin(), m_virtualSessionId.end());
     data.willMessage->qos = 2;
     data.willMessage->topic = TOPIC_WILLMESSAGE;
     data.willMessage->retain = false;
-    data.willMessage->delayInterval = m_sessionExpiryInterval + 1000;   // the sessionExpiryInterval has priority
+    data.willMessage->delayInterval = m_sessionExpiryInterval + 1000; // the sessionExpiryInterval has priority
 
     m_client->startConnection(connection, data);
-    m_client->subscribe(connection, { {{m_virtualSessionId + "/#", 2, false, true, 2}} });
-    m_client->subscribe(connection, { {{TOPIC_WILLMESSAGE, 2, false, true, 2}} });
+    m_client->subscribe(connection, {{{m_virtualSessionId + "/#", 2, false, true, 2}}});
+    m_client->subscribe(connection, {{{TOPIC_WILLMESSAGE, 2, false, true, 2}}});
 
     std::unique_lock<std::mutex> lock(m_mutex);
     m_connection = connection;
-
 }
 
 IStreamConnectionPtr ProtocolMqtt5Client::getConnection() const
@@ -249,7 +239,7 @@ void ProtocolMqtt5Client::disconnect()
     message->addSendPayload(m_virtualSessionId);
     m_client->publish(connection, std::move(dataWill), message);
 
-    IMqtt5Client::DisconnectData data = { ReasonCodeDisconnectWithWillMessage, "", {} };
+    IMqtt5Client::DisconnectData data = {ReasonCodeDisconnectWithWillMessage, "", {}};
     m_client->endConnection(connection, data);
 
     if (connection)
@@ -263,37 +253,37 @@ std::uint32_t ProtocolMqtt5Client::getProtocolId() const
     return PROTOCOL_ID;
 }
 
-bool ProtocolMqtt5Client::areMessagesResendable() const 
+bool ProtocolMqtt5Client::areMessagesResendable() const
 {
     return false;
 }
 
-bool ProtocolMqtt5Client::doesSupportMetainfo() const 
+bool ProtocolMqtt5Client::doesSupportMetainfo() const
 {
     return true;
 }
 
-bool ProtocolMqtt5Client::doesSupportSession() const 
+bool ProtocolMqtt5Client::doesSupportSession() const
 {
     return true;
 }
 
-bool ProtocolMqtt5Client::needsReply() const 
+bool ProtocolMqtt5Client::needsReply() const
 {
     return false;
 }
 
-bool ProtocolMqtt5Client::isMultiConnectionSession() const 
+bool ProtocolMqtt5Client::isMultiConnectionSession() const
 {
     return false;
 }
 
-bool ProtocolMqtt5Client::isSendRequestByPoll() const 
+bool ProtocolMqtt5Client::isSendRequestByPoll() const
 {
     return false;
 }
 
-bool ProtocolMqtt5Client::doesSupportFileTransfer() const 
+bool ProtocolMqtt5Client::doesSupportFileTransfer() const
 {
     return false;
 }
@@ -366,17 +356,17 @@ void ProtocolMqtt5Client::sendMessage(IMessagePtr message)
     }
 }
 
-void ProtocolMqtt5Client::moveOldProtocolState(IProtocol& /*protocolOld*/) 
+void ProtocolMqtt5Client::moveOldProtocolState(IProtocol& /*protocolOld*/)
 {
 }
 
-bool ProtocolMqtt5Client::received(const IStreamConnectionPtr& connection, const SocketPtr& socket, int bytesToRead) 
+bool ProtocolMqtt5Client::received(const IStreamConnectionPtr& connection, const SocketPtr& socket, int bytesToRead)
 {
     bool ok = m_client->receive(connection, socket, bytesToRead);
     return ok;
 }
 
-hybrid_ptr<IStreamConnectionCallback> ProtocolMqtt5Client::connected(const IStreamConnectionPtr& /*connection*/) 
+hybrid_ptr<IStreamConnectionCallback> ProtocolMqtt5Client::connected(const IStreamConnectionPtr& /*connection*/)
 {
     std::unique_lock<std::mutex> lock(m_mutex);
     auto callback = m_callback.lock();
@@ -402,7 +392,7 @@ void ProtocolMqtt5Client::disconnected(const IStreamConnectionPtr& /*connection*
     }
 }
 
-IMessagePtr ProtocolMqtt5Client::pollReply(std::deque<IMessagePtr>&& /*messages*/) 
+IMessagePtr ProtocolMqtt5Client::pollReply(std::deque<IMessagePtr>&& /*messages*/)
 {
     return {};
 }
@@ -419,7 +409,7 @@ void ProtocolMqtt5Client::subscribe(const std::vector<std::string>& subscribtion
     {
         const std::string& subscription = subscribtions[i];
         std::string topic = "/" + subscription;
-        data.subscriptions.push_back({ topic, 2, false, false, 2 });
+        data.subscriptions.push_back({topic, 2, false, false, 2});
     }
 
     std::unique_lock<std::mutex> lock(m_mutex);
@@ -444,7 +434,6 @@ void ProtocolMqtt5Client::cycleTime()
     }
 }
 
-
 // IMqtt5ClientCallback
 void ProtocolMqtt5Client::receivedConnAck(const ConnAckData& data)
 {
@@ -466,7 +455,7 @@ void ProtocolMqtt5Client::receivedConnAck(const ConnAckData& data)
                 m_sessionExpiryInterval = data.sessionExpiryInterval;
             }
         }
-//        IStreamConnectionPtr connection = m_connection;
+        //        IStreamConnectionPtr connection = m_connection;
         auto callback = m_callback.lock();
         lock.unlock();
 
@@ -489,7 +478,7 @@ void ProtocolMqtt5Client::receivedConnAck(const ConnAckData& data)
     }
     else
     {
-        streamError << "Connection failed, reason code: " << (unsigned int)data.reasoncode << " (" << data.reasonString << ")";
+        streamError << "Connection failed, reason code: " << static_cast<unsigned int>(data.reasoncode) << " (" << data.reasonString << ")";
     }
 }
 
@@ -531,12 +520,10 @@ void ProtocolMqtt5Client::receivedPublish(const PublishData& data, const IMessag
 
 void ProtocolMqtt5Client::receivedSubAck(const std::vector<std::uint8_t>& /*reasoncodes*/)
 {
-
 }
 
 void ProtocolMqtt5Client::receivedUnsubAck(const std::vector<std::uint8_t>& /*reasoncodes*/)
 {
-
 }
 
 void ProtocolMqtt5Client::receivedPingResp()
@@ -552,12 +539,11 @@ void ProtocolMqtt5Client::receivedPingResp()
 
 void ProtocolMqtt5Client::receivedDisconnect(const DisconnectData& data)
 {
-    streamInfo << "Disconnect from mqtt5 broker. Reason Code: " << (unsigned int)data.reasoncode << " (" << data.reasonString << ")";
+    streamInfo << "Disconnect from mqtt5 broker. Reason Code: " << static_cast<unsigned int>(data.reasoncode) << " (" << data.reasonString << ")";
 }
 
 void ProtocolMqtt5Client::receivedAuth(const AuthData& /*data*/)
 {
-
 }
 
 void ProtocolMqtt5Client::closeConnection()
@@ -574,8 +560,6 @@ void ProtocolMqtt5Client::setProtocolSessionData(const IProtocolSessionDataPtr& 
 {
 }
 
-
-
 //---------------------------------------
 // ProtocolMqtt5Factory
 //---------------------------------------
@@ -588,13 +572,10 @@ struct RegisterProtocolMqtt5ClientFactory
     }
 } g_registerProtocolMqtt5ClientFactory;
 
-
-
-
 // IProtocolFactory
 IProtocolPtr ProtocolMqtt5ClientFactory::createProtocol(const Variant& data)
 {
     return std::make_shared<ProtocolMqtt5Client>(data);
 }
 
-}   // namespace finalmq
+} // namespace finalmq
